+++ pocketsmith/api/__init__.py
@@ -8,7 +8,8 @@ from pocketsmith.api.attachments_api import AttachmentsApi
 from pocketsmith.api.budgeting_api import BudgetingApi
 from pocketsmith.api.categories_api import CategoriesApi
 from pocketsmith.api.category_rules_api import CategoryRulesApi
-from pocketsmith.api.data_connections_api import DataConnectionsApi
+from pocketsmith.api.data_feeds_accounts_api import DataFeedsAccountsApi
+from pocketsmith.api.data_feeds_connections_api import DataFeedsConnectionsApi
 from pocketsmith.api.institutions_api import InstitutionsApi
 from pocketsmith.api.transaction_accounts_api import TransactionAccountsApi
 from pocketsmith.api.transactions_api import TransactionsApi